requirements.txt
boto3==1.40.5
botocore==1.40.5
jmespath==1.0.1
joblib==1.5.1
numpy==1.26.4
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.13.1
six==1.16.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.2.1
xgboost==3.0.3
